further folder reorganization; mainly capital-case convention with components
