+++ app/App.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import React, { Component } from 'react';
-import TabBar from './components/TabBar.js';
+import TabBar from './components/TabBar/TabBar.js';
 
 class App extends Component {
   constructor(props) {
