+++ app/components/Chart/Chart.js
@@ -12,8 +12,8 @@ import {
   TouchableOpacity,
   ActivityIndicator
 } from 'react-native';
-import styles from './styles';
-import { getDbData, formatElapTime } from '../utils';
+import styles from '../styles';
+import { getDbData, formatElapTime } from '../../../utils';
 
 var ds = new ListView.DataSource({rowHasChanged: (r1, r2) => r1 !== r2});
 var greatestPercentage = 0;
