+++ app/components/HomeButton/HomeButton.js
@@ -12,7 +12,7 @@ import {
   AlertIOS,
   View
 } from 'react-native';
-import styles from './styles';
+import styles from '../styles';
 import {
   getCurrentLocation,
   getAddress,
@@ -22,9 +22,9 @@ import {
   addUpdateTime,
   nearbySearch,
   formatToTime
-} from '../utils';
-import {seed} from '../database';
-import {SERVER_ROUTE} from '../server/env/development';
+} from '../../../utils';
+import {seed} from '../../../database';
+import {SERVER_ROUTE} from '../../../server/env/development';
 
 console.disableYellowBox = true;
