+++ app/components/TabBar/TabBar.js
@@ -7,19 +7,19 @@ import {
   Text,
   View
 } from 'react-native'
-import styles from './styles';
+import styles from '../styles';
 
 import {
   HomeRender,
   MapRender,
   LogRender,
   ChartRender
-} from './compiledRender';
+} from '../compiledRender';
 
-var homeIcon = require('../../resources/home.png');
-var mapIcon = require('../../resources/compass.png')
-var logIcon = require('../../resources/diary.png');
-var chartIcon = require('../../resources/pie-chart.png');
+var homeIcon = require('../../../resources/home.png');
+var mapIcon = require('../../../resources/compass.png')
+var logIcon = require('../../../resources/diary.png');
+var chartIcon = require('../../../resources/pie-chart.png');
 
 class TabBar extends Component {
   constructor(props) {
